Refactored API functionality into helper files
