+++ .eslintrc
@@ -8,7 +8,8 @@
     "indent": ["error", 2]
   },
   "parserOptions": {
-    "ecmaVersion": 8
+    "ecmaVersion": 8,
+    "sourceType": "module"
   },
   "env": {
     "node": true,
